requirements.txt
pandas
numpy==1.24.0
scikit-learn
nltk
regex
transformers==4.28.1
datasets==2.12.0
pandas==1.5.3
openai
torch==2.6.0
python-dotenv
google-cloud-pubsub
chardet
XlsxWriter
Flask==2.2.2
werkzeug==2.2.2
google-cloud-pubsub==2.10.0
SQLAlchemy==2.0.12